requirements.txt
aioauth-client==0.10.0
aiohttp==2.3.2
aiohttp-jinja2==0.14.0
async-timeout==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
colorclass==2.2.0
dateutils==0.6.6
docopt==0.6.2
dynaconf==0.5.2
idna==2.6
Jinja2==2.9.6
markdown2==2.3.4
MarkupSafe==1.0
Mastodon.py==1.1.1
multidict==3.3.2
nltk==3.2.5
packaging==16.8
pip-upgrader==1.4.3
py==1.4.34
PyJWT==1.5.3
pyparsing==2.2.0
pytest==3.2.3
pytest-aiohttp==0.1.3
python-box==3.1.1
python-dateutil==2.6.1
python-engineio==1.7.0
python-socketio==1.8.1
pytz==2017.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
terminaltables==3.1.0
urllib3==1.22
xxhash==1.0.1
yarl==0.13.0
